refactor: move song_writer and MidiSong to song module
