+++ bevy_midix/src/synth/mod.rs
@@ -2,10 +2,13 @@
 Synthesizer resources, setup and plugins
 "#]
 
-use crate::prelude::SoundFont;
+use crate::{
+    prelude::SoundFont,
+    song::{SongId, SongWriter},
+};
 use bevy::prelude::*;
 use crossbeam_channel::Sender;
-use midix::prelude::{ChannelVoiceMessage, Timed};
+use midix::prelude::ChannelVoiceMessage;
 use std::sync::Mutex;
 use thiserror::Error;
 use tinyaudio::OutputDevice;
@@ -152,42 +155,5 @@ impl Default for Synth {
             synthesizer: SynthState::NotLoaded,
             _device: None,
         }
-    }
-}
-
-/// This defines a song, a file, or otherwise
-/// that has timestamps associated with midi events.
-///
-/// this is named as such not to conflict with [`midix::MidiSource`]
-pub trait SongWriter {
-    /// Create sink commands this type.
-    fn song_id(&self) -> Option<SongId> {
-        None
-    }
-    /// A list of timed events relevant to this song
-    fn events(&self) -> impl Iterator<Item = Timed<ChannelVoiceMessage>>;
-    /// is this song looped?
-    fn looped(&self) -> bool {
-        false
-    }
-    // /// Helper method that will divide events into individual channels
-    // fn divide_events_into_channels(&self) -> FnvHashMap<Channel, Vec<Timed<ChannelVoiceMessage>>> {
-    //     let mut map = FnvHashMap::default();
-
-    //     for event in self.events() {}
-
-    //     todo!()
-    // }
-}
-
-impl SongWriter for MidiSong {
-    fn song_id(&self) -> Option<SongId> {
-        Some(self.id)
-    }
-    fn events(&self) -> impl Iterator<Item = Timed<ChannelVoiceMessage>> {
-        self.events.iter().copied()
-    }
-    fn looped(&self) -> bool {
-        self.looped
     }
 }
